# More performant DotNetEventListener.OnEventWritten