Hi

Since this method is called a lot, I figured this small performance improvement would be worth doing :-)

| Method |     Mean |     Error |    StdDev | Rank |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |---------:|----------:|----------:|-----:|-------:|------:|------:|----------:|
|  Before | 6.373 ns | 0.2041 ns | 0.3681 ns |    2 | 0.0096 |     - |     - |      40 B |
| After | 1.886 ns | 0.0604 ns | 0.0536 ns |    1 |      - |     - |     - |         - |