@djluck 2019-11-05
Definitely, appreciate the PR mate. I'm trying to work out some benchmarking code to help evaluate performance changes like this. Once that work is done, I'll merge this change.