@bdrupieski 2019-11-13
@blankensteiner Just curious how did you benchmark it and create that table?