@blankensteiner 2019-11-13
@bdrupieski I used https://benchmarkdotnet.org/