@blankensteiner 2019-11-13
I ran the tests with these attributes: [RPlotExporter, RankColumn, MemoryDiagnoser]
This will give you a table (in the console) and when copy/pasting it to the GitHub markdown editor, it will format it nicely :-)