@djluck 2020-01-30
@blankensteiner Just to confirm, did you have `Configuration=Release` set for the build when you ran these benchmarks?