@blankensteiner 2020-01-30
Hi @djluck Yes I was running a release build.